Teatime V1.0
Teatime is now a fully functional timer.
The program will check the terminal size before and during execution to prevent "cursing" the program.
The time is now limited to 99:59 or 5999 seconds.
The timer can be stopped and resumed with 'p' key.
Once the infusion is done, Teatime will stay on screen until any key is pressed.